% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{confidence}
\alias{confidence}
\title{Common argument}
\arguments{
\item{confidence}{A vector of percentiles giving the coverage of confidence intervals.
It must be greater than 0 and less than 1.
If \code{NULL}, no confidence intervals are produced.}
}
\description{
Common argument
}
\keyword{internal}

