% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{control}
\alias{control}
\title{Common argument}
\arguments{
\item{control}{Passed directly to \code{\link{rq.fit.sfn}()} during Robust STR decomposition.}
}
\description{
Common argument
}
\keyword{internal}

