% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{gapCV}
\alias{gapCV}
\title{Common argument}
\arguments{
\item{gapCV}{An optional parameter defining the length of the sequence of
skipped values in the cross validation procedure.}
}
\description{
Common argument
}
\keyword{internal}
