% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grocery.R
\docType{data}
\name{grocery}
\alias{grocery}
\title{Grocery and supermarkets turnover}
\format{An object of class \code{ts}.}
\usage{
grocery
}
\description{
Turnover of supermarkets and grocery stores in New South Wales, Australia.
}
\examples{
plot(grocery, ylab = "NSW Grocery, $ 10^6")
}
\references{
Australian Bureau of Statistics, CAT 8501.0.
(\href{http://tinyurl.com/hvx2vx2}{TABLE 11. Retail Turnover, State by Industry Subgroup, 
Original})
}
\keyword{datasets}

