% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{lambdas}
\alias{lambdas}
\title{Common argument}
\arguments{
\item{lambdas}{An optional parameter.
A structure which replaces lambda parameters provided with predictors.
It is used as either a starting point for the optimisation of parameters or as the
exact model parameters.}
}
\description{
Common argument
}
\keyword{internal}

