% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{nMCIter}
\alias{nMCIter}
\title{Common argument}
\arguments{
\item{nMCIter}{Number of Monte Carlo iterations used to estimate confidence intervals for Robust STR decomposition.}
}
\description{
Common argument
}
\keyword{internal}

