% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{pattern}
\alias{pattern}
\title{Common argument}
\arguments{
\item{pattern}{An optional parameter which has the same structure as \code{lambdas}
although with a different meaning. All zero values correspond to lambda
(smoothing) parameters which will not be estimated.}
}
\description{
Common argument
}
\keyword{internal}
