% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBeta.R
\name{plotBeta}
\alias{plotBeta}
\title{Plots the varying beta coefficients of decomposition}
\usage{
plotBeta(x, xTime = NULL, predictorN = 1, dim = c(1, 2, 3), type = "o",
  pch = 20, palette = function(n) rainbow(n, start = 0, end = 0.7))
}
\arguments{
\item{x}{Result of STR decomposition.}

\item{xTime}{Times for data to plot.}

\item{predictorN}{Predictor number in the decomposition to plot the corresponding beta coefficiets.}

\item{dim}{Dimensions to use to plot the beta coefficients.
When \code{1}, the standard charts are used.
When \code{2}, \code{graphics:::filled.contour} function is used.
When \code{3}, \code{rgl:::persp3d} is used. The default value is \code{1}.}

\item{type}{Type of the graph for one dimensional plots.}

\item{pch}{Symbol code to plot points in 1-dimensional charts. Default value is \code{20}.}

\item{palette}{Color palette for 2 - and 3 - dimentional plots.}
}
\description{
\code{plotBeta} plots the varying beta coefficients of STR decomposition.
It plots coefficients only only for independent seasons (one less season than defined).
}
\examples{
\dontrun{

fit <- AutoSTR(log(grocery))
for(i in 1:2) plotBeta(fit, predictorN = i, dim = 2)
}
}
\author{
Alexander Dokumentov
}
\seealso{
\code{\link{plot.STR}}
}

