% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{reltol}
\alias{reltol}
\title{Common argument}
\arguments{
\item{reltol}{An optional parameter which is passed directly to \code{\link{optim}()}
when optimising the parameters of the model.}
}
\description{
Common argument
}
\keyword{internal}

