% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{reportDimensionsOnly}
\alias{reportDimensionsOnly}
\title{Common argument}
\arguments{
\item{reportDimensionsOnly}{A boolean parameter. When TRUE the method constructs the
design matrix and reports its dimensions without proceeding further.
It is mostly used for debugging.}
}
\description{
Common argument
}
\keyword{internal}
