% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{robust}
\alias{robust}
\title{Common argument}
\arguments{
\item{robust}{When \code{TRUE}, Robust STR decomposition is used. Default is \code{FALSE}.}
}
\description{
Common argument
}
\keyword{internal}
