% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasadj.R
\name{seasadj.STR}
\alias{seasadj.STR}
\title{Seasonal adjustment}
\usage{
seasadj.STR(object)
}
\arguments{
\item{object}{Result of STR decomposition.}
}
\description{
\code{seasadj.STR} extracts seasonally adjusted data by removing the seasonal components from the result of STR decomposition.
}
\examples{
\dontrun{

fit <- AutoSTR(log(grocery))
plot(seasadj(fit))

}
}
\author{
Alexander Dokumentov
}
\seealso{
\code{\link{STRmodel}}, \code{\link{RSTRmodel}}, \code{\link{STR}}, \code{\link{AutoSTR}}
}

