% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{solver}
\alias{solver}
\title{Common argument}
\arguments{
\item{solver}{A vector with two string values. Possible combinations are:
c("MatrixModels", "cholesky") (default), c("MatrixModels", "qr"),
c("Matrix", "cholesky"), c("Matrix", "qr").
The parameter is used to specify a particular library
and method to solve the minimisation problem during STR decompositon.}
}
\description{
Common argument
}
\keyword{internal}

