% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{strDesign}
\alias{strDesign}
\title{Common argument}
\arguments{
\item{strDesign}{An optional parameter used to create the design
matrix. It is used internally in the library to improve performance when the design
matrix does not require full recalculation.}
}
\description{
Common argument
}
\keyword{internal}

