% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls.R
\docType{data}
\name{calls}
\alias{calls}
\title{Number of phone calls dataset}
\format{A numerical time series of class \code{msts} and \code{ts}.}
\source{
\href{http://robjhyndman.com/data/callcenter.txt}{Data file}
}
\usage{
calls
}
\description{
Number of call arrivals per 5-minute interval handled on weekdays between 7:00 am and 9:05 pm
from March 3, 2003 in a large North American commercial bank.
}
\examples{
plot(calls, ylab = "Calls handled")
}
\references{
Forecasting Time Series With Complex Seasonal Patterns Using Exponential Smoothing
Alysha M. De Livera, Rob J. Hyndman & Ralph D. Snyder
(\href{http://www.tandfonline.com/doi/pdf/10.1198/jasa.2011.tm09771}{Journal of the 
American Statistical Association})
}
\keyword{datasets}

