% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{components}
\alias{components}
\title{Extract STR components}
\usage{
components(object)
}
\arguments{
\item{object}{Result of STR decomposition.}
}
\description{
\code{components} extracts components as time series from the result of an STR decomposition.
}
\examples{
\dontrun{

fit <- AutoSTR(log(grocery))
comp <- components(fit)
plot(comp)
}
}
\author{
Alexander Dokumentov
}
\seealso{
\code{\link{STRmodel}}, \code{\link{RSTRmodel}}, \code{\link{STR}}, \code{\link{AutoSTR}}
}

