% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{solver}
\alias{solver}
\title{Common argument}
\arguments{
\item{solver}{is "MatrixModels" or "cholesky". Used to specify a particular library
and method to solve the minimisation problem during STR decompositon.}
}
\description{
Common argument
}
\keyword{internal}

